+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added inventor names to raw csv benchmarkd datasets (useful for validation and debugging).
+- Added cluster precision and cluster recall estimators (for cluster block sampling).
 - Added additional inventor benchmark datasets.
 
 ## 0.0.1 (June 27, 2022)